HW2/requirements.txt
boto3
flask
xxhash
jump-consistent-hash
jmespath